+++ tests/testthat/test-labels.R
@@ -4,7 +4,7 @@ test_that("default assignment", {
   x1 <- remove_labels(x)
 
   expect_false(inherits(x, "labelled")) # Hmisc::label produces this class
-  expect_equal(attr(x, "label"), "runs")
+  expect_equal(exattr(x, "label"), "runs")
 
   expect_error(assign_labels(x, NULL))
   expect_error(assign_labels(x, 1:2))
@@ -87,4 +87,9 @@ test_that("view_labels() works", {
 
   View <- NA
   expect_error(view_labels(df), "Something went wrong")
+})
+
+test_that("exact match [141]", {
+  x <- struct(1L, "integer", labels = 1:2)
+  expect_identical(get_labels(x), NA_character_)
 })